+++ src/geoai3d/io/__init__.py
@@ -8,15 +8,18 @@ reference system through the round trip.
 """
 
 from geoai3d.io.dispatch import read_lidar
+from geoai3d.io.geotiff import read_geotiff, to_geotiff
 from geoai3d.io.las import read_las, to_las
 from geoai3d.io.parquet import read_parquet, to_parquet
 from geoai3d.io.xyz import read_xyz, to_xyz
 
 __all__ = [
+    "read_geotiff",
     "read_las",
     "read_lidar",
     "read_parquet",
     "read_xyz",
+    "to_geotiff",
     "to_las",
     "to_parquet",
     "to_xyz",